move opt_result
